+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-12-08
+### Added
+- Added commands for moving the player and for following another multibox player
+### Fixed
+- Fixed a bug where Inspire command would only Inspire one multibox fighter
+
 ## [1.0.0] - 2019-11-29
 ### Added
 - :tada: Initial commit!
 
-[1.0.0]: https://github.com/elodon/TkMemory/commit/master
+[1.1.0]: https://github.com/elodon/TkMemory/commit/master
+[1.0.0]: https://github.com/elodon/TkMemory/commit/e4043a44b6a79430a1307f0ef8b0a5ddaf068999